+++ src/gs/gsColor.cpp
@@ -2,7 +2,7 @@
 #include "gsMath.h"
 #include <ctime>
 
-gs::RandomRange<uint8_t> gs::Color::randByte(0, 255, time(0));
+gs::RandomRange<uint8_t> gs::Color::randByte(0, 255, (int) time(0));
 
 gs::Color::Color()
     :   gs::Vec3<uint8_t>(randByte.Sample(), randByte.Sample(), randByte.Sample()),
@@ -29,7 +29,7 @@ gs::Color::Color(const gs::Vec3i& rhs)
 }
 
 gs::Color::Color(const gs::Vec3d& rhs)
-	: gs::Vec3<uint8_t>((uint8_t) gs::Clamp(rhs.x, 0.0, 255.0), (uint8_t) gs::Clamp(rhs.y, 0.0, 255.0), gs::Clamp(rhs.z, 0.0, 255.0)),
+	: gs::Vec3<uint8_t>((uint8_t) gs::Clamp(rhs.x, 0.0, 255.0), (uint8_t) gs::Clamp(rhs.y, 0.0, 255.0), (uint8_t) gs::Clamp(rhs.z, 0.0, 255.0)),
 	  r(x), g(y), b(z)
 {
 }
